requirements.txt
matplotlib==3.8.3
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
PyYAML==6.0.1
torch==2.2.1
torchvision==0.17.1
torchmetrics
